some more channel testing

+++ lib/nacs-seq/awg/DummyClient.h
@@ -56,11 +56,11 @@ public:
         Phase
     };
     struct ChannelInfo {
-        uint32_t m_phys_chn;
+        uint8_t m_phys_chn;
         uint32_t m_chn_num;
         uint32_t linear_idx = 0; // Linear channel number for BasicSeq's list of pulses
         ChnType m_chn_type;
-        ChannelInfo(uint32_t phys_chn, uint32_t chn_num, ChnType chn_type)
+        ChannelInfo(uint8_t phys_chn, uint32_t chn_num, ChnType chn_type)
             : m_phys_chn(phys_chn),
               m_chn_num(chn_num),
               m_chn_type(chn_type)
